pyproject.toml
#:schema https://json.schemastore.org/pyproject.json
[project]
name = "tso_api"
version = "0.1.0"
description = "TSO API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]==0.115.12",
    "httpx[h2,brotli]==0.28.1",
    "pillow==11.2.1",
    "psycopg[binary,pool]==3.2.9",
    "pydantic==2.11.4",
    "pydantic-settings==2.9.1",
    "pyhumps==3.8.0",
    "pyjwt[crypto]==2.10.1",
    "python-multipart==0.0.20",
    "recipe-scrapers==15.7.1",
    "uuid6==2024.7.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
asyncio_default_fixture_loop_scope = "function"
asyncio_default_test_loop_scope = "function"
asyncio_mode = "auto"
markers = ['collections']

[tool.ruff]
line-length = 120
required-version = ">=0.9.6"
src = ["src", "test"]

[tool.ruff.format]
quote-style = "single"
skip-magic-trailing-comma = true

[tool.ruff.lint]
preview = true
select = ["ALL"]
ignore = [
  # conflict with formatter
  "W191",
  "E111",
  "E114",
  "E117",
  "D206",
  "D300",
  "Q000",
  "Q001",
  "Q002",
  "Q003",
  "COM812",
  "COM819",
  "ISC002",
  "E501",
  "D203",
  "D212",
  "D205",
  "DOC201",
  "FBT001",
  # maybe use these at some point
  "D100",
  "D101",
  "D102",
  "D103",
  "D104",
  "D105",
  "D106",
  "D107",
  "CPY001",
  "S404",
  "S603",
  "ANN202",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "TRY002", "TRY003", "EM101", "FBT003", "S311", "PLR2004"]
"!src/tso_api/routers/*.py" = ["ANN201"]
"src/tso_api/models/*.py" = ["RUF012"]


[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = false

[tool.ruff.lint.mccabe]
max-complexity = 50

[dependency-groups]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.26.0",
    "pytest-cov==6.1.1",
    "pytest-subtests==0.14.1",
    "ruff==0.11.10",
    "testcontainers[postgres]==4.10.0",
]
